requirements.txt
# Flask Blog Platform Dependencies

# Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1

# Database
SQLAlchemy==2.0.23

# Forms and Validation
WTForms==3.1.1
email-validator==2.1.0

# Security
Werkzeug==3.0.1

# Utilities
python-slugify==8.0.1
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
